Move the responsibility on folder replacement to `GradleProject`.
